:truck: Move env-based configuration validators to own module
